refactor copy_config import
